fix tests and utils for assignment

+++ src/main/java/org/learn/coursera/outline/week8/twosum/TwoSumFinder.java
@@ -4,5 +4,5 @@ import java.util.List;
 
 public interface TwoSumFinder {
 
-    Integer find(final List<Integer> numbers, final Integer sum);
+    Integer find(final List<Long> numbers, final Integer sum);
 }
